backend/requirements.txt
Flask
python-dotenv
Flask-PyMongo
oauthlib
requests
requests-oauthlib
authlib